requirements.txt
flask
requests
beautifulsoup4
mysql-connector-python
pytz
yfinance
pandas
flasgger
